Move FeedbackMailer outside PA/Nsm namespaces
Its shared by the app generally, although we may
need to customize the application name used at some
point.
